ElasticNet implementation to solve regression problem and compare it with LASSO regression and Ridge regression.
Problem : As Elastic-Net is generally implemented datasets containing 1000 or more parameters, instead of a
a given problem, let us assume and generate random values to be used in the model. These values will be divided into training and testing data and then
we will compute the differences in the co-efficients and errors between
Ridge, LASSO and Elastic-Net.